# openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: 'http://localhost'
    description: Generated server url
paths:
  '/api/teams/{id}':
    get:
      tags:
        - team-controller
      summary: Получить команду по ID
      operationId: getTeamById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Успешно
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '401':
          description: Пользователь не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Команда не найдена
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
      security:
        - BearerAuth: []
    put:
      tags:
        - team-controller
      summary: Обновить команду
      operationId: updateTeam
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTeamRequest'
        required: true
      responses:
        '200':
          description: Команда обновлена
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '401':
          description: Пользователь не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Команда не найдена
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '409':
          description: Имя команды занято
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
      security:
        - BearerAuth: []
    delete:
      tags:
        - team-controller
      summary: Удалить команду
      operationId: deleteTeam
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Команда удалена
        '401':
          description: Пользователь не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Команда не найдена
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
      security:
        - BearerAuth: []
...
  /api/teams:
    get:
      tags:
        - team-controller
      summary: Получить все команды
      operationId: getAllTeams
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum:
              - ACTIVE
              - INACTIVE
      responses:
        '200':
          description: Успешно
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeamResponse'
        '401':
          description: Пользователь не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
      security:
        - BearerAuth: []
    post:
      tags:
        - team-controller
      summary: Создание команды
      operationId: createTeam
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeamRequest'
        required: true
      responses:
        '201':
          description: Команда создана
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '400':
          description: Ошибка валидации
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Пользователь не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '409':
          description: Команда с таким именем уже существует
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
      security:
        - BearerAuth: []
  '/api/teams/{id}/members':
    get:
      tags:
        - team-controller
      summary: Получить участников команды
      operationId: getTeamMembers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Успешно
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeamResponse'
        '401':
          description: Пользователь не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: Нет доступа к команде
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
      security:
        - BearerAuth: []
    post:
      tags:
        - team-controller
      summary: Добавить участника в команду
      operationId: addTeamMember
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddTeamMemberRequest'
        required: true
      responses:
        '200':
          description: Успешно
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '401':
          description: Пользователь не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: Недостаточно прав
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Команда или пользователь не найдены
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
      security:
        - BearerAuth: []
...
  /api/teams/my-teams:
    get:
      tags:
        - team-controller
      summary: Получить команды текущего пользователя
      operationId: getMyTeams
      responses:
        '200':
          description: Успешно
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeamResponse'
        '401':
          description: Пользователь не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MessageResponse'
      security:
        - BearerAuth: []
...
components:
  schemas:
    UpdateTeamRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
          minLength: 3
        status:
          type: string
          enum:
            - ACTIVE
            - INACTIVE
      required:
        - name
    MessageResponse:
      type: object
      properties:
        message:
          type: string
    TeamResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum:
            - ACTIVE
            - INACTIVE
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateTeamRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
          minLength: 3
      required:
        - name
    AddTeamMemberRequest:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        role:
          type: string
          enum:
            - ADMIN
            - EDITOR
            - VIEWER
      required:
        - role
        - userId
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT